small nit refactorss; RenameAtomicUtils to be in services;
